+++ tests/run-pass/fn/onearg.sk
